third_party/WebKit/LayoutTests/http/tests/security/suborigins/crossorigin/suborigin-cors-fetch-failure-output-expected.txt

CONSOLE ERROR: Failed to load http://127.0.0.1:8000/security/resources/cors-script.php?cors=false: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
CONSOLE ERROR: Failed to load http://127.0.0.1:8000/security/resources/cors-script.php?cors=false: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
This is a testharness.js-based test.
PASS Custom headers causes preflight failure 
PASS Lack of Access-Control-Allow-Suborigin on response causes failure 
Harness: the test ran to completion.
